Changements après les remarques
